finally working for the AWS scope... still a ton of work!
