+++ package.json
@@ -26,6 +26,12 @@
             "extensions": [".nf"],
             "firstLine": "^#!.*\\bnextflow\\b",
             "configuration": "./language-configuration.json"
+        },{
+            "id": "nextflow-config",
+            "aliases": ["Nextflow-config", "nextflow-config"],
+            "extensions": [".config"],
+            "firstLine": "^#!.*\\bnextflow\\b",
+            "configuration": "./language-configuration.json"
         }],
         "grammars": [{
             "language": "nextflow",
@@ -41,8 +47,35 @@
                 "language": "nextflow",
                 "path": "./snippets/snippets.json"
             }
-        ]
+        ],
+        "configuration": {
+			"type": "object",
+			"title": "Example configuration",
+			"properties": {
+				"languageServerExample.maxNumberOfProblems": {
+					"scope": "resource",
+					"type": "number",
+					"default": 100,
+					"description": "Controls the maximum number of problems produced by the server."
+				},
+				"languageServerExample.trace.server": {
+					"scope": "window",
+					"type": "string",
+					"enum": [
+						"off",
+						"messages",
+						"verbose"
+					],
+					"default": "off",
+					"description": "Traces the communication between VS Code and the language server."
+				}
+			}
+		}
     },
+    "activationEvents": [
+		"onLanguage:nextflow-config"
+	],
+	"main": "./client/out/extension",
     "scripts": {
 		"vscode:prepublish": "npm run compile",
 		"compile": "tsc -b",
